+++ src/arm32/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2021-2024 Gabriel Bjørnager Jensen.
+// Copyright 2024 Gabriel Bjørnager Jensen.
 //
 // This file is part of Pollex.
 //
@@ -24,7 +24,12 @@
 //! This includes T variants of Arm32.
 
 use crate::use_mod;
-use_mod!(pub address);
-use_mod!(pub condition);
+use_mod!(pub arm_opcode);
+use_mod!(pub predicate);
+use_mod!(pub flag);
 use_mod!(pub instruction);
 use_mod!(pub register);
+use_mod!(pub shifter);
+use_mod!(pub signed);
+use_mod!(pub thumb_opcode);
+use_mod!(pub unsigned);